Cleanup greedy module
These cosmetic changes make things a bit more idiomatic. In Erlang, `case`
is almost always favored over `if`. If you don't have another standard, use
4 spaces per indent. Try to keep lines under 80 characters long. Always keep
code under 120 characters. Establish a line length standard and stick to it.
You can bump guard clauses to the next line, particularly for a function
signature.

This code may not compile. I'd test it, but the code in the repo is already
busted.