Remove generated edocs
This patch removes the `/doc` directory from the project. There is no need
to store generated documentation. If you did, you'd need to run the
`make edoc` target before every commit to make sure the docs were up-to-date
with the source code.

Just generate the docs locally from the `Makefile`.